+++ ChangeLog.md
@@ -5,6 +5,11 @@ XP Framework Core ChangeLog
 
 ## 7.4.0 / 2016-06-04
 
+### Bugfixes
+
+* Fixed passing unicode arguments via `Runtime::newInstance()`
+  (@thekid)
+
 ### Features
 
 * Extended `lang.FunctionType`'s constructor to accept type names in